refactor : 각 파일의 목적에 맞게 파일명 변경
