+++ app/src/main/java/com/minionz/qrna/view/ui/dashboard/DashboardViewModel.kt
@@ -6,8 +6,4 @@ import androidx.lifecycle.ViewModel
 
 class DashboardViewModel : ViewModel() {
 
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
-    }
-    val text: LiveData<String> = _text
 }
