+++ app/src/main/java/com/minionz/qrna/view/ui/home/HomeViewModel.kt
@@ -6,8 +6,4 @@ import androidx.lifecycle.ViewModel
 
 class HomeViewModel : ViewModel() {
 
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
-    }
-    val text: LiveData<String> = _text
 }
