+++ app/src/main/res/values/strings.xml
@@ -10,7 +10,7 @@
     <string name="sign_up_confirm">확인</string>
     <string name="sign_up_cancel">취소</string>
     <string name="title_activity_main">MainActivity</string>
-    <string name="title_home">Home</string>
+    <string name="title_home">홈</string>
     <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="title_my_info">내 정보</string>
 </resources>
